refactor(ble): extract Web Bluetooth transport into @boardsesh/ble-protocol
The Expo-web mobile adapter copied the Web Bluetooth transport helpers
(characteristic probing, chunked write series, request-device options)
verbatim from packages/web. Behavioural hotfixes — the NotSupportedError
with-response retry, the MoonBoard RedBearLab fallback — would have to
land in two places and drift.

Move the pure, browser-side logic into a shared, renderer-agnostic
@boardsesh/ble-protocol/web-transport module and import it from both the
Next.js web adapter and the mobile web adapter. The module declares its
own minimal Web Bluetooth type surface (a structural subset of the W3C
spec) and reaches navigator.bluetooth through globalThis, so it needs no
lib.dom or @types/web-bluetooth — keeping it out of the mobile OTA
fingerprint. web's bluetooth-shared / bluetooth-aurora / bluetooth-moonboard
now re-export from the shared module so existing web imports keep resolving.

This also removes the mobile-local web-bluetooth.d.ts, whose unscoped
global Navigator/BluetoothDevice declarations previously applied to the
entire mobile TS project, native files included.
