--- packages/mobile/src/lib/ble/web-bluetooth.d.ts
@@ -1,63 +1,0 @@
-// Ambient Web Bluetooth DOM types for the Expo-web BLE adapter.
-//
-// The native fingerprint graph must not gain a `@types/web-bluetooth`
-// devDependency (it would change bun.lock / packages/mobile/package.json and so
-// perturb the OTA fingerprint), so the minimal surface the web adapter needs is
-// declared locally instead. Only the members `web-adapter.ts` touches are
-// modelled — this is not the full W3C Web Bluetooth spec. lib.dom.d.ts ships
-// none of these, so there is no declaration-merge conflict (verified against
-// TypeScript 6.0's lib.dom).
-
-type BluetoothServiceUUID = number | string;
-type BluetoothCharacteristicUUID = number | string;
-
-interface BluetoothLEScanFilter {
-  services?: BluetoothServiceUUID[];
-  name?: string;
-  namePrefix?: string;
-}
-
-interface RequestDeviceOptions {
-  filters?: BluetoothLEScanFilter[];
-  optionalServices?: BluetoothServiceUUID[];
-  acceptAllDevices?: boolean;
-}
-
-interface BluetoothCharacteristicProperties {
-  readonly write: boolean;
-  readonly writeWithoutResponse: boolean;
-  readonly read: boolean;
-  readonly notify: boolean;
-}
-
-interface BluetoothRemoteGATTCharacteristic {
-  readonly properties: BluetoothCharacteristicProperties;
-  writeValueWithResponse(value: BufferSource): Promise<void>;
-  writeValueWithoutResponse(value: BufferSource): Promise<void>;
-}
-
-interface BluetoothRemoteGATTService {
-  getCharacteristic(characteristic: BluetoothCharacteristicUUID): Promise<BluetoothRemoteGATTCharacteristic>;
-}
-
-interface BluetoothRemoteGATTServer {
-  readonly connected: boolean;
-  connect(): Promise<BluetoothRemoteGATTServer>;
-  disconnect(): void;
-  getPrimaryService(service: BluetoothServiceUUID): Promise<BluetoothRemoteGATTService>;
-}
-
-interface BluetoothDevice extends EventTarget {
-  readonly id: string;
-  readonly name?: string;
-  readonly gatt?: BluetoothRemoteGATTServer;
-}
-
-interface Bluetooth {
-  getAvailability(): Promise<boolean>;
-  requestDevice(options?: RequestDeviceOptions): Promise<BluetoothDevice>;
-}
-
-interface Navigator {
-  readonly bluetooth?: Bluetooth;
-}
